remove duplicate methods
